5.1.6

See merge request BI/magreport-internal-2.1!84

+++ magreport-backend/src/main/java/ru/magnit/magreportbackend/metrics_function/impl/SumIntegerFunction.java
@@ -4,12 +4,12 @@ import ru.magnit.magreportbackend.metrics_function.MetricsFunction;
 
 public class SumIntegerFunction implements MetricsFunction {
 
-    private int sumOfValues;
+    private long sumOfValues;
 
     @Override
     public void addValue(String value, int col, int row) {
         if (value != null)
-            sumOfValues += Double.parseDouble(value);
+            sumOfValues += Long.parseLong(value);
     }
 
     @Override
